Changes Proposal to make use of extracted components
